clippy: -D unnecessary_cast

+++ capsules/extra/src/can.rs
@@ -518,7 +518,7 @@ impl<'a, Can: can::Can> can::ReceiveClient<{ can::STANDARD_CAN_PACKET_SIZE }>
                                 up_calls::UPCALL_MESSAGE_RECEIVED,
                                 (
                                     0,
-                                    shared_len as usize,
+                                    shared_len,
                                     match id {
                                         can::Id::Standard(u16) => u16 as usize,
                                         can::Id::Extended(u32) => u32 as usize,
