+++ capsules/extra/src/proximity.rs
@@ -270,9 +270,7 @@ impl hil::sensors::ProximityClient for ProximitySensor<'_> {
                     if app.enqueued_command_type == ProximityCommand::ReadProximityOnInterrupt {
                         // Case: ReadProximityOnInterrupt
                         // Only callback to those apps which we expect would want to know about this threshold reading.
-                        if ((temp_val as u8) > app.upper_proximity)
-                            || ((temp_val as u8) < app.lower_proximity)
-                        {
+                        if (temp_val > app.upper_proximity) || (temp_val < app.lower_proximity) {
                             upcalls.schedule_upcall(0, (temp_val as usize, 0, 0)).ok();
                             app.subscribed = false; // dequeue
                         }
