+++ libraries/tock-cells/src/numeric_cell_ext.rs
@@ -56,22 +56,22 @@ where
     }
 
     fn increment(&self) {
-        self.set(self.get() + T::from(1 as usize));
+        self.set(self.get() + T::from(1_usize));
     }
 
     fn decrement(&self) {
-        self.set(self.get() - T::from(1 as usize));
+        self.set(self.get() - T::from(1_usize));
     }
 
     fn get_and_increment(&self) -> T {
         let ret = self.get();
-        self.set(ret + T::from(1 as usize));
+        self.set(ret + T::from(1_usize));
         ret
     }
 
     fn get_and_decrement(&self) -> T {
         let ret = self.get();
-        self.set(ret - T::from(1 as usize));
+        self.set(ret - T::from(1_usize));
         ret
     }
 }
